EditorState爆誕
