--- src/Editor.tsx
@@ -1,49 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { useTranslation } from "react-i18next";
-import Forms from "./Forms";
-import "./Editor.css";
-
-type Props = {
-  lang: string;
-  owner: string | null;
-  repo: string | null;
-  autoFill: boolean;
-};
-
-export default function Editor(props: Props) {
-  const { lang, owner, repo, autoFill } = props;
-  const { t } = useTranslation();
-  const [showAlert, setShowAlert] = useState(false);
-
-  return (
-    <div className="editor">
-      <div className="edit-area">
-        <h2>Input Form</h2>
-        <div className="edit-inner">
-          <form>
-            <Forms />
-            <p className="proviso">
-              {"If you generate a README, you are deemed to have accepted the "}
-              <a href="?scene=document&pageid=terms-of-service">Terms of Service</a>
-              {". "}
-              {
-                "And if you upload README that is generated by LEADYOU to your public repository on GitHub, it will be on "
-              }
-              <a href="/">New Generated READMEs</a>
-              {". "}
-              {"If you do not want it, please refer "}
-              <a href="?scene=document&pageid=help">Help</a>
-              {" page."}
-            </p>
-            <input type="submit" value="Generate" />
-            {showAlert && <p className="fill-alert"></p>}
-          </form>
-        </div>
-      </div>
-      <div className="preview-area">
-        <h2>Preview README</h2>
-        <div className="md-content"></div>
-      </div>
-    </div>
-  );
-}
